refactor: move ChakraProvider to its own package
