+++ packages/provider/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@chakra-ui/react",
+  "name": "@chakra-ui/provider",
   "version": "1.6.6",
-  "description": "Responsive and accessible React UI components built with React and Emotion",
+  "description": "Theme and components agnostic ChakraProvider",
   "main": "dist/cjs/index.js",
   "module": "dist/esm/index.js",
   "types": "dist/types/index.d.ts",
@@ -34,53 +34,10 @@
     "lint:types": "tsc --noEmit"
   },
   "dependencies": {
-    "@chakra-ui/accordion": "1.3.5",
-    "@chakra-ui/alert": "1.2.7",
-    "@chakra-ui/avatar": "1.2.8",
-    "@chakra-ui/breadcrumb": "1.2.8",
-    "@chakra-ui/button": "1.4.2",
-    "@chakra-ui/checkbox": "1.5.5",
-    "@chakra-ui/close-button": "1.1.11",
-    "@chakra-ui/control-box": "1.0.15",
-    "@chakra-ui/counter": "1.1.8",
     "@chakra-ui/css-reset": "1.0.0",
-    "@chakra-ui/editable": "1.2.8",
-    "@chakra-ui/form-control": "1.4.0",
     "@chakra-ui/hooks": "1.5.5",
-    "@chakra-ui/icon": "1.1.11",
-    "@chakra-ui/image": "1.0.18",
-    "@chakra-ui/input": "1.2.9",
-    "@chakra-ui/layout": "1.4.8",
-    "@chakra-ui/live-region": "1.0.14",
-    "@chakra-ui/media-query": "1.1.2",
-    "@chakra-ui/menu": "1.7.2",
-    "@chakra-ui/modal": "1.8.10",
-    "@chakra-ui/number-input": "1.2.9",
-    "@chakra-ui/pin-input": "1.6.4",
-    "@chakra-ui/popover": "1.8.2",
-    "@chakra-ui/popper": "2.2.1",
-    "@chakra-ui/portal": "1.2.8",
-    "@chakra-ui/progress": "1.1.12",
-    "@chakra-ui/provider": "1.6.6",
-    "@chakra-ui/radio": "1.3.9",
-    "@chakra-ui/react-env": "1.0.6",
-    "@chakra-ui/select": "1.1.13",
-    "@chakra-ui/skeleton": "1.1.17",
-    "@chakra-ui/slider": "1.2.8",
-    "@chakra-ui/spinner": "1.1.12",
-    "@chakra-ui/stat": "1.1.12",
-    "@chakra-ui/switch": "1.2.8",
     "@chakra-ui/system": "1.7.2",
-    "@chakra-ui/table": "1.2.6",
-    "@chakra-ui/tabs": "1.5.4",
-    "@chakra-ui/tag": "1.1.12",
-    "@chakra-ui/textarea": "1.1.13",
-    "@chakra-ui/theme": "1.10.0",
-    "@chakra-ui/toast": "1.2.10",
-    "@chakra-ui/tooltip": "1.3.9",
-    "@chakra-ui/transition": "1.3.4",
-    "@chakra-ui/utils": "1.8.2",
-    "@chakra-ui/visually-hidden": "1.0.14"
+    "@chakra-ui/utils": "1.8.2"
   },
   "peerDependencies": {
     "@emotion/react": "^11.0.0",
