+++ packages/provider/src/index.tsx
@@ -1,0 +1,1 @@
+export * from "./chakra-provider"
